requirements.txt
fastapi
uvicorn
numpy
pillow
streamlit
python-multipart